refactor(*): rename all Optional types
Move 'Optional' from the beginning of the names
to the end.
e.g. instead of OptionalButtonProps we now have
ButtonPropsOptional. Event though that's not the
correct english form (the old form is), this new form
with Optional at the end of the term makes it easier
to work on code, specially because we have a lint
rule to have imports in alphabetical order, making both
imports (Optional and normal) close together.

e.g. instead of having this:

import {
  Button,
  ButtonProps,
  ButtonVariant,
  ColorTheme,
  InteractionType,
  OptionalButtonProps,
} from '@reflex-ui/core';

We have this:

import {
  Button,
  ButtonProps,
  ButtonPropsOptional,
  ButtonVariant,
  ColorTheme,
  InteractionType,
} from '@reflex-ui/core';