feat(scrollable) add scrollable function for web app scroll
When building web apps with flexbox and scrolling content,
we should disable HTML's body scroll add height: 100% to
a few DOM nodes, so we can use div-based scrolling
like the ones when using <ScrollView> and <FlatList>.
This is what this commit does via scrollable() function.
When its called, it adds those CSS styles to the <head>
of the HTML document, getting it ready for div-based scrolling.
This commit also changes showcase app to call this function
on App.web.tsx.